requirements.txt
fastapi
uvicorn
python-multipart
mercadopago
google-cloud-storage
python-dotenv
requests
openai
python-docx
